Updated the edit and delete actions. Now only authors can delete theirs articles.

+++ app/Http/Controllers/ArticlesController.php
@@ -5,13 +5,14 @@ namespace App\Http\Controllers;
 use App\Article;
 use App\Post;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class ArticlesController extends Controller
 {
-    public function __construct()
-    {
-        $this->middleware('auth');
-    }
+//    public function __construct()
+//    {
+//        $this->middleware('auth');
+//    }
 
     /**
      * Display a listing of the resource.
@@ -73,6 +74,10 @@ class ArticlesController extends Controller
      */
     public function edit(Article $article)
     {
+        if ($article->user->id != Auth::id()) {
+            abort(403, 'You cannot edit an article which is not written by you!');
+        }
+
         return view('articles.edit', compact('article'));
     }
 
@@ -85,6 +90,10 @@ class ArticlesController extends Controller
      */
     public function update(Request $request, Article $article)
     {
+        if ($article->user->id != Auth::id()) {
+            abort(403, 'You cannot edit an article which is not written by you!');
+        }
+
         $article->title = $request->input('title');
         $article->content = $request->input('content');
         try {
@@ -104,6 +113,10 @@ class ArticlesController extends Controller
      */
     public function destroy(Article $article)
     {
+        if ($article->user->id != Auth::id()) {
+            abort(403, 'You cannot delete an article which is not written by you!');
+        }
+
         try {
             $article->delete();
         } catch (\Exception $e) {
